Move erc20 utils into lib
