+++ sewup/src/token/erc20.rs
@@ -1,8 +1,8 @@
-use ewasm_api::types::*;
-use sewup::token::helpers::{
+use super::helpers::{
     copy_into_array, copy_into_storage_value, get_allowance, get_balance, set_allowance,
     set_balance,
 };
+use ewasm_api::types::{Address, StorageValue};
 
 pub fn do_transfer(recipient: Address, value: StorageValue) {
     let sender = ewasm_api::caller();
