+++ sewup/src/token/mod.rs
@@ -5,3 +5,9 @@ mod tests;
 
 #[cfg(not(test))]
 pub mod helpers;
+
+// This will be test after compiled into wasm
+#[cfg(not(test))]
+pub mod erc20;
+
+pub mod signature;
